+++ RELEASE_NOTES.rst
@@ -33,6 +33,7 @@ describe future plans.
     ~~~~~
 
     * Honour ``r1`` / ``r2`` arguments in ``AdHocSolver.calculate_UB``.  :issue:`56`
+    * Honour ``r1`` / ``r2`` arguments in ``DiffcalcSolver.calculate_UB``.  :issue:`58`
 
 0.3.1
 ######